+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - [TD-1533] Added structure schema type for system params
 - [TD-1690] index implementation results on elastic
 
+### Removed
+
+- [TD-1902] Removed description field from RuleImplementation
+
 ### Changed
 
 - [TD-1893] Use CI_JOB_ID instead of CI_PIPELINE_ID